backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Backend + Browser Automation Template"

requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "dishka>=1.7.2",
    "fastapi[all]>=0.119.1",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "taskiq>=0.11.18",
    "taskiq-redis>=1.1.2",
    "yarl>=1.22.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "httpx[http2]>=0.28.1",
]

[project.optional-dependencies]
emulation = [
    "playwright>=1.58.0",
    "fake-useragent>=2.2.0",
    "google-generativeai>=0.8.0",
]

[project.scripts]
app = "app:main"
cli = "cli:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "coverage>=7.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.0",
    "ruff>=0.14.1",
    "pre-commit>=3.7.0",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"

[tool.ruff]
target-version = "py313"
line-length = 88
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
lint.ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
